Annotating RSEM gene.results using ENSEMBL gtf and refGenome CRAN package

This function executes the docker container annotate.1, where refGenome is used to annotated gene.results and isoforms.results outputs from RSEM using ENSEMBL GTF annotation

annoByGtfchipseq(
  group = "docker",
  peaks.file = getwd(),
  gtf.file,
  extension = 10000
)

Arguments

group,

a character string. Two options: "sudo" or "docker", depending to which group the user belongs

peaks.file,

a character string indicating the MACS peak file, extension _peaks.xls, with full path

gtf.file,

a character string indicating the file, with full path for the genome gtf

extension,

a number defining how many nucleotides should be expandend the extremes of the targetr gene to find an overlap with peaks, default 10000

Value

one file: MACS2 peaks annotated

Author

Raffaele Calogero

Examples

if (FALSE) {
    #downloading fastq files
    #running rsemannoByGtfchipseq
    annoByGtfchipseq(group="docker", peaks.file=paste(getwd(),"h3k9me1_igg_peaks.xls", sep="/"), 
               gtf.file=paste("/Users/raffaelecalogero/Dropbox/courses/DUKENUS_JUL2019/course/course/datasets/genomes/mm10bwa","Mus_musculus.GRCm38.97.gtf", sep="/"), extension=10000)
}