Generating bowtie2 genome index

This function executes the docker container bowtie2 where bowtie2-2.2.9 is installed. The index is created using ENSEMBL genome fasta file. User needs to provide the URL for ENSEMBL genome located in the ENSEMBL ftp

bowtie2Index(
  group = c("sudo", "docker"),
  genome.folder = getwd(),
  ensembl.urlgenome = NULL,
  ensembl.urlgtf = NULL,
  threads = 1
)

Arguments

group,

a character string. Two options: "sudo" or "docker", depending to which group the user belongs

genome.folder,

a character string indicating the folder where the indexed reference genome for Bowtie will be located

ensembl.urlgenome,

a character string indicating the URL from ENSEMBL ftp for the unmasked genome sequence of interest

ensembl.urlgtf,

a character string indicating the URL from ENSEMBL ftp for the GTF for genome of interest

threads,

a number indicating the number of cores to be used from the application

Value

The indexed Bowtie genome reference sequence

Author

Raffaele Calogero

Examples

if (FALSE) {
    #running rsemstar index for human
    bowtie2Index(group="docker",genome.folder="/data/scratch/hg38bowtie2",
    ensembl.urlgenome=
    "ftp://ftp.ensembl.org/pub/release-87/fasta/homo_sapiens/dna/Homo_sapiens.GRCh38.dna.toplevel.fa.gz",
    ensembl.urlgtf=
    "ftp://ftp.ensembl.org/pub/release-87/gtf/homo_sapiens/Homo_sapiens.GRCh38.87.gtf.gz",
    threads=8)

}