Running CircHunter circRNA quantification module

This function executes the docker container circhunter by running the circRNA quantification module of CircHunter to quantify the level of expression of a set of circRNA BS sequences in a given RNA-Seq experiment. For CircHunter algorithm detail please refer to: https://github.com/carlo-deintinis/circhunter/tree/master/CircHunter.

circrnaQuantification(
  group = c("sudo", "docker"),
  scratch.folder,
  rnaseq.data,
  backsplicing_junctions.data,
  hc.params
)

Arguments

group,
a character string. Two options: "sudo" or "docker", depending to which group the user belongs

scratch.folder,
a character string indicating the scratch folder where docker container will be mounted

rnaseq.data,
string indicating the path to the fastq file of the RNA-Seq dataset to analyse

backsplicing_junctions.data,
string indicating the path to the fasta file of the circRNA back-splicing sequences to search in the RNA-Seq dataset

hc.params,
vector of six parameters to set the analysis. The element of the vector indicate in order: the k-mer size, the thread number, the dimension of the hash table, the dimension of the collision list, the number of k-mers that must be matched to the sequence to consider the sequence itself as represented in the RNA-Seq data, and the number of perfect matches required in the k-mer to consider it matched to a sequence.

Value

A count table reporting the number of RNA-Seq reads supporting specific circRNA back-splice junctions

Examples

if (FALSE) {

    #retrieve the example data
    system("wget https://github.com/carlo-deintinis/circhunter/archive/master.zip") #'     system("unzip master.zip")
    system("gzip ./circhunter-master/CircHunter/data/test_rna-seq.fastq.gz")

    #running the circrnaQuantification function
    circrnaQuantification(group="docker", scratch.folder="/data/scratch", rnaseq.data=paste(getwd,"/circhunter-master/CircHunter/data/test_rna-seq.fastq.gz", sep=""), backsplicing_junctions.data=paste(getwd,"/circhunter-master/CircHunter/data/CRC_circRNA_backsplicing_sequences.fasta", sep=""), hc.params=c(27, 40, 1000000, 1000000, 17, 30))

}