Running CIRI v2 tool for circRNAs prediction

This function executes the docker container ciri2 where CIRI v2.0.6 is installed and it provides the list of circRNAs predicted from a RNA-Seq experiment. For CIRI 2 tool detail refer to: "Gao, Y., Zhang, J., & Zhao, F. (2017). Circular RNA identification based on multiple seed matching. Brief Bioinform. 2018 Sep 28;19(5):803-810."

ciri2(
  group = c("sudo", "docker"),
  scratch.folder,
  sam.file,
  genome.file,
  annotation.file = NA,
  max.span = 2e+05,
  stringency.value = c("high", "low", "zero"),
  quality.threshold = 10,
  threads = 1
)

Arguments

group,
a character string. Two options: "sudo" or "docker", depending to which group the user belongs

scratch.folder,
a character string indicating the scratch folder where docker container will be mounted

sam.file,
a character string indicating the path to the RNA-Seq alignment SAM file from BWA

genome.file,
a character string indicating the path to the Fasta file of the reference genomic sequence (it should be the same reference indexed for the BWA alignment)

annotation.file,
a character string indicating the path to the GTF/GFF file reporting the reference gene annotations

max.span,
an integer reporting the maximum spanning distance of a circRNA (default = 200000 bp)

stringency.value,
the selected stringency level of the analysis. Three possible options are available: "high" (high stringency, default), in which CIRI2 only provides circRNAs supported by more than 2 distinct Paired Chiastic Clipping (PCC) signals; "low" (low stringency), CIRI2 only provides circRNAs supported by more than 2 junction reads; "zero", CIRI2 provides all circRNAs regardless junction read counts or PCC signals

quality.threshold,
integer indicating the threshold for mapping quality of each segment of junction reads (default=10)

threads,
integer indicating the number of threads used for the analysis (default=1)

Value

The list of CIRI 2 predicted circRNAs

Examples

if (FALSE) {

    #retrieve the example data
    system("wget https://sourceforge.net/projects/ciri/files/CIRI2/CIRI_v2.0.6.zip") #retrieve the example data
    system("unzip CIRI_v2.0.6.zip")

    #running ciri2 function
    ciri2(group="docker", scratch.folder="/data/scratch", sam.file=paste(getwd(),"/CIRI_v2.0.6/data/sample.sam", sep=""), genome.file=paste(getwd(),"/CIRI_v2.0.6/data/chr1.fa", sep=""), annotation.file="", max.span=200000, stringency.value="high", quality.threshold=10, threads=1)
}