hfc, heatmap for clustering

This function generate heatmap and other plot based on clustering and on a specific gene list

heatmapBase(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  status = 0,
  lower.range = -1,
  upper.range = 1
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

scratch.folder,

a character string indicating the path of the scratch folder

file,

a character string indicating the path of the file, with counts.table name and extension included

status,

0 if is raw count, 1 otherwise

lower.range,

the lower range of signal in the heatmap.

upper.range,

the upper range of signal in the heatmap.

Value

A heatmap.

Author

Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Examples

if (FALSE) {
system("wget http://130.192.119.59/public/heatmap_test.zip")
system("unzip heatmap_test.zip")
setwd("heatmap_test")
heatmapBase(group="docker",scratch.folder="/data/scratch",file=paste(getwd(),"DEfiltered__log2TPM.txt", sep="/"), status=1, lower.range=-1, upper.range=1)
}