generating a PCA from counts, FPKM and TPM tables from rnaseqCounts outuputs

This function generates PCA plot from counts, FPKM and TPM tables from rnaseqCounts outuputs.

pca(
  experiment.table = "./_counts.txt",
  type = c("counts", "FPKM", "TPM"),
  covariatesInNames = FALSE,
  samplesName = TRUE,
  principal.components = c(1, 2),
  legend.position = c("bottom", "bottomleft", "left", "topleft", "top", "topright",
    "right", "center"),
  pdf = TRUE,
  output.folder = getwd()
)

Arguments

experiment.table,

a character string indicating the counts, FPKM or TPM table file name and its path

type,

a character value indicating the content of the file: counts, FPKM or TPM

covariatesInNames,

a boolean value indicating if covariates are inserted after \_ in the filename

samplesName,

a boolean value indicating if in the plot samples names are plotted or not

principal.components,

a numerical vector with two values indicating the principal components to be plotted

legend.position,

a character string indicating the location of the covariates legend

pdf,

a boolean value indicating if results has to be saved in a pdf

output.folder,

output folder

Value

Returns a PCA plot

Author

Raffaele Calogero

Examples

if (FALSE) {
  system("wget 130.192.119.59/public/test.analysis.zip")
  unzip("test.analysis.zip")
  setwd("test.analysis")
  library(docker4seq)
  pca(experiment.table="./_log2FPKM.txt", type="FPKM",
      legend.position="topleft", covariatesInNames=FALSE, samplesName=TRUE,
      principal.components=c(1,2), pdf = TRUE,
      output.folder=getwd())

}