Platypus filter analysis NOT READY TO GO ON STABLE to be used in case the full platypus workflow fail.

This function executes platypusFilter filtering the single_variants.vcf generated by platypys function.

platypusFilter(
  group = c("sudo", "docker"),
  data.folder = getwd(),
  scratch.folder,
  GQ,
  minSampGQ,
  NR,
  minSampNR,
  NV,
  minSampNV,
  normal_samples,
  GT_normal,
  minSampGT_normal,
  tumoral_samples,
  GT_tumoral,
  minSampGT_tumoral,
  stringent_filter,
  annotation
)

Arguments

group,

a character string. Two options: sudo or docker, depending to which group the user belongs

data.folder,

a character string indicating the folder where bams and vcf files are located and where output will be written

scratch.folder,

a character string indicating the path of the scratch folder

GQ,

min GQ value to consider (extreme included)

minSampGQ,

min number of samples with GQ value (extreme included), usually we start with 85% of the samples

NR,

min NR value to consider (extreme included), the number of reads covering the SNV region

minSampNR,

min number of samples with NR value (extreme included), usually we start with 85% of the samples

NV,

min NV value to consider (extreme included), the n of reads with the SNV

minSampNV,

min number of samples with NV value (extreme included)

normal_samples,

string with names (group names of bam files) of normal samples separated by hash, write NULL if you do not want to use GT filter in normal samples

GT_normal,

GT value in normal samples to consider, type "NO" if you do not want to use this filter, else e.g. you might use 0/0

minSampGT_normal,

min number of normal samples with GT value

tumoral_samples,

string with names (group names of bam files) of tumoral samples separated by &, write NULL if you do not want to use GT filter in normal samples

GT_tumoral,

GT value in tumoral samples to consider, type "NO" if you do not want to use this filter. Type e.g. 0/0 if you do not want to consider this genotype.

minSampGT_tumoral,

min number of tumoral samples in which the GT value is NOT present

stringent_filter,

To enable the filter (it keeps only the variants with "PASS" value or the variants that have only the "alleleBias" value) use 1 or 0 to disable it

annotation,

hg19 and mm10 are actually available for the annotation of the detected SNVs

Author

Riccardo Panero, riccardo.panero[at]gmail[dot]com, Bioinformatics and Genomics unit, University of Torino Italy

Examples

if (FALSE) {
    #filtering platypus results 
 platypusFilter(group="docker", data.folder="/archive/home/rcaloger/data/platypus_tests/mm10ENSEMBL", 
        scratch.folder="/scratch/users/rcaloger/",
        GQ=10, minSampGQ=2, NR=10, minSampNR=2, NV=3, 
        minSampNV=1, normal_samples="MAMBO43", GT_normal="0/0", 
        minSampGT_normal=1, tumoral_samples="MAMBO43TRT#MAMBO43TRNT", GT_tumoral="0/0", 
        minSampGT_tumoral=1, stringent_filter=0, annotation="mm10")
}