Running RNAseq counting workflow for a single sample

This function executes a set of docker containers allowing the generation of gene and isoforms counts for a single sample. #params skewer

rnaseqCounts(
  group = "sudo",
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  threads = 4,
  adapter5,
  adapter3,
  seq.type = "pe",
  min.length = 40,
  genome.folder = "/data/genomes/hg38star",
  strandness = "none",
  save.bam = TRUE,
  org = "hg38",
  annotation.type = "gtfENSEMBL"
)

Arguments

group,

a character string. Two options: "sudo" or "docker", depending to which group the user belongs

fastq.folder,

a character string indicating where gzip fastq files are located

scratch.folder,

a character string indicating the scratch folder where docker container will be mounted

threads,

a number indicating the number of cores to be used from the application

adapter5,

a character string indicating the fwd adapter

adapter3,

a character string indicating the rev adapter

seq.type,

a character string indicating the type of reads to be generated by the sequencer. Two options: "se" or "pe" respectively for single end and pair end sequencing.

min.length,

a number indicating minimal length required to return a trimmed read #params rsemstar

genome.folder,

a character string indicating the folder where the indexed reference genome is located. IMPORTANT the present function only suport genomic indexes made using ensembl genom and the corresponding gtf

strandness,

a character string indicating the type ofsequencing protocol used for the analysis. Three options: "none", "forward", "reverse" respectively for non strand selection, reverse for Illumina strandness protocols, reverse for ACCESS Illumina protocol

save.bam,

a boolean value, TRUE or FALSE, to save also BAM files generated by STAR and RSEM #params rsemanno

org,

a character string indicating the genome assembly used for mapping and counting with "rsemstar" function only required for biocENSEMBL based annotation

annotation.type,

a character string. Two options: "biocENSEMBL" or "gtfENSEMBL". "biocENSEMBL" will annotate by Bioconductor only protein coding genes. "gtfENSEMBL" will annotate all RNAs described in "annotation.type"

Value

Returns the output of skewer, rsemstar, rsemannos' functions

Author

Raffaele Calogero

Examples

if (FALSE) {
system("wget http://130.192.119.59/public/test_R1.fastq.gz")
library(docker4seq)
rnaseqCounts(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch/",
            adapter5="AGATCGGAAGAGCACACGTCTGAACTCCAGTCA",
            adapter3="AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT",
            seq.type="se", threads=24,  min.length=40,
            genome.folder="/data/genomes/hg38star", strandness="none", save.bam=FALSE,
            org="hg38", annotation.type="gtfENSEMBL")
}