Running small RNA-seq single-end reads alignment and quantification using BWA and custom scripts

This function executes the docker container where BWA is installed. BWA is a read alignment package that efficiently align short sequencing reads against a large reference sequence. Alignment is performed against annotations of human small RNAs. Read count is performed by GenomicAlignments R package and custom Python and bash commands.

sncRNA(
  group = c("sudo", "docker"),
  fastq.folder = getwd(),
  scratch.folder,
  mode,
  reference,
  threads = 1,
  mb.url.haripin = NULL,
  mb.url.mature = NULL,
  mb.species = NULL,
  adapter.type = c("NEB", "ILLUMINA", "QIAGEN", "LEXOGEN", "DIAGENE", "SEQMATIC",
    "TRILINK"),
  trimmed.fastq = FALSE
)

Arguments

group,

a character string. Two options: "sudo" or "docker", depending to which group the user belongs

fastq.folder,

a character string indicating where trimmed fastq files are located

scratch.folder,

a character string indicating the scratch folder where docker container will be mounted

mode,

a character string indicating the required type of analysis. Compatible analyses mode are "miRNA" and "ncRNA". In "miRNA" analysis mode, the version ("mb.version" argument) and species prefix ("mb.species" argument) of miRBase are required. This mode require also the "reference" argument. In the "ncRNA" mode only the "reference" argument is required.

reference,

a character string indicating the path to the reference fasta file used to create the BWA index

threads,

a number indicating the number of cores to be used from the application

mb.url.haripin,

character string indicating the link to the hairpin miRNA sequences miRBase database. Visit http://www.mirbase.org to select the proper version number.

mb.url.mature,

a character string indicating the link to the mature miRNA sequences from miRBase database. Visit http://www.mirbase.org to select the proper version number.

mb.species,

a character string indicating the three-letter prefix of a species annotated in miRBase (e.g. "hsa" for human miRNAs). Please refer to http://www.mirbase.org/help/genome_summary.shtml to obtain the proper species prefix.

adapter.type,

a character string. Seven options are available depending on which miRNA library prep was used: NEB, ILLUMINA, QIAGEN, LEXOGEN, DIAGENE, SEQMATIC, TRILINK

trimmed.fastq,

a boolean logical variable indicating if trimmed fastq are saved. Default is FALSE

Value

Read count table of RNA-Seq reads aligned miRNA or non-miRNA annotations

Author

Giulio Ferrero

Examples